fix: Fix positioning reference output
In the case of having 3+ heads, positioning one of them (target)
relative to another (reference), potentially the reference is moved from
its original position. This is bad because if you previously positioned
the reference relative to this third head, you will loose this setup.

In short, now the positioning commands are associative. Meaning, that a
series of them is eventually guaranteed to reach the globally desired
layout.